Add color nav
