+++ src/components/pages/Homepage.js
@@ -1,4 +1,4 @@
-import { Nav } from "./Navbar";
+import { Nav } from "../layouts/Nav";
 
 export default function Homepage() {
   return (
